Permit trailing commas in macro invocations
This commit uses the `?` operator to support optional trailing commas in
macro calls without requiring the call to use the panic-message arms.

The `?` macro operator was stabilized in `1.32` for the 2018 edition,
and `1.37` for the 2015 edition. Applying this commit requires raising
the crate MSRV to one of these two versions; for the lower, the crate
must be explicitly set to the `edition = "2018"` key in its manifest.